avatar-user-photo
- add change of the avatar, when user is login
- add css
