+++ src/components/Layout/Layout.jsx
@@ -1,20 +1,7 @@
 import { Outlet } from 'react-router-dom';
-
+import HeaderMarkup from 'components/HeaderMarkup/HeaderMarkup.jsx';
 // import { Suspense } from 'react';
-import {
-  Header,
-  LogoSvg,
-  UserAvatar,
-  Nav,
-  StyledLinkLogo,
-  StyledLink,
-  Container,
-  LeftNavWrapper,
-  NavMenu,
-  SettingsIcon,
-  LogOutIcon,
-  NavLi
-} from './Layout.styled.js';
+import { Header } from './Layout.styled.js';
 import { useState } from 'react';
 import { Paskal } from 'components/pascal/paskal.jsx';
 import { TeamModal } from 'components/teams/TeamModal.jsx';
@@ -27,37 +14,7 @@ export const Layout = () => {
   return (
     <>
       <Header>
-        <Container>
-          <Nav>
-            <li>
-              <StyledLinkLogo to="/">
-                <LogoSvg alt="logo" />
-                Tracker <br /> of water
-              </StyledLinkLogo>
-            </li>
-
-            <li style={{width:"118px"}}>
-              <LeftNavWrapper>
-                <StyledLink to="/signup" style={{justifyContent: "flex-end"}}>
-                  Sign in
-                  <UserAvatar alt="user avatar" />
-                </StyledLink>
-
-                <NavMenu className="navMenu">
-                  <NavLi>
-                    <SettingsIcon/>
-                    <a href="/settings" style={{color: "#407bff"}}>Settings</a>
-                  </NavLi>
-                  <NavLi>
-                    <LogOutIcon/>
-                    <a href="/logout" style={{color: "#407bff"}}>Log out</a>
-                  </NavLi>
-                </NavMenu>
-
-              </LeftNavWrapper>
-            </li>
-          </Nav>
-        </Container>
+        <HeaderMarkup />
       </Header>
       <main>
         <Outlet />
