--- src/components/Layout/Select.styled.js
@@ -1,21 +1,0 @@
-// import styled from 'styled-components';
-
-// export const NavMenu = styled.ul`
-//   /* display: none; */
-//   position: absolute;
-//   z-index: 1;
-//     &:hover{
-//         display: block
-//     }
-//   /* 
-
-//     opacity: 0;
-//     transform: translateY(-10px);
-//     visibility: hidden;
-//     transition: 0.3s;
-//     &:active{
-//         transform: translateY(0);
-//         opacity: 1;
-//         visibility: visible;
-//     } */
-// `;
