+++ src/components/redux/auth/operations.js
@@ -1,10 +1,10 @@
-import axios from "axios";
-import { createAsyncThunk } from "@reduxjs/toolkit";
+import axios from 'axios';
+import { createAsyncThunk } from '@reduxjs/toolkit';
 
-axios.defaults.baseURL = "http://localhost:8000/api";
+axios.defaults.baseURL = 'http://localhost:8000/api';
 
 const setAuthHeader = token => {
-    axios.defaults.headers.common.Authorization = `Bearer ${token}`;
+  axios.defaults.headers.common.Authorization = `Bearer ${token}`;
 };
 
 /*const clearAuthHeader = () => {
@@ -12,47 +12,46 @@ const setAuthHeader = token => {
 };*/
 
 export const signUp = createAsyncThunk(
-    "auth/signup",
-    async (credentials, thunkAPI) => {
-        try {
-            const res = await axios.post("/users/register", credentials);
-            setAuthHeader(res.data.token);
-            return res.data;
-        } catch (error) {
-            return thunkAPI.rejectWithValue(error.message);
-        }
+  'auth/signup',
+  async (credentials, thunkAPI) => {
+    try {
+      const res = await axios.post('/users/register', credentials);
+      setAuthHeader(res.data.token);
+      return res.data;
+    } catch (error) {
+      return thunkAPI.rejectWithValue(error.message);
     }
+  }
 );
 
 export const signIn = createAsyncThunk(
-    "auth/signin",
-    async (credentials, thunkAPI) => {
-        try {
-            const res = await axios.post("/users/login", credentials);
-            setAuthHeader(res.data.token);
-            return res.data;
-        } catch (error) {
-            return thunkAPI.rejectWithValue(error.message);
-        }
+  'auth/signin',
+  async (credentials, thunkAPI) => {
+    try {
+      const res = await axios.post('/users/login', credentials);
+      setAuthHeader(res.data.token);
+      return res.data;
+    } catch (error) {
+      return thunkAPI.rejectWithValue(error.message);
     }
+  }
 );
 
-
 export const refreshUser = createAsyncThunk(
-    "auth/refresh",
-    async (_, thunkAPI) => {
-        const state = thunkAPI.getState();
-        const persistedToken = state.auth.token;
+  'auth/refresh',
+  async (_, thunkAPI) => {
+    const state = thunkAPI.getState();
+    const persistedToken = state.auth.token;
 
-        if (persistedToken === null) {
-            return thunkAPI.rejectWithValue("Unable to fetch user");
-        }
-        try {
-            setAuthHeader(persistedToken);
-            const res = await axios.get("/users/current");
-            return res.data;
-        } catch (error) {
-            return thunkAPI.rejectWithValue(error.message);
-        }
+    if (persistedToken === null) {
+      return thunkAPI.rejectWithValue('Unable to fetch user');
     }
+    try {
+      setAuthHeader(persistedToken);
+      const res = await axios.get('/users/current');
+      return res.data;
+    } catch (error) {
+      return thunkAPI.rejectWithValue(error.message);
+    }
+  }
 );
