+++ src/components/redux/auth/selectors.js
@@ -3,3 +3,5 @@ export const selectIsLoggedIn = state => state.auth.isLoggedIn;
 export const selectUser = state => state.auth.user;
 
 export const selectIsRefreshing = state => state.auth.isRefreshing;
+
+export const selectIcon = state => state.auth.icon;
